requirements.txt
openai
requests
python-dotenv
schedule
crontab
snowflake-connector-python